backend/requirements.txt
Collecting pylint-django
  Downloading https://files.pythonhosted.org/packages/40/e7/6276ed3bc88500c553d98d745fdca2685683c769051ba208e1d66f5007c1/pylint_django-2.0.13-py3-none-any.whl (69kB)
Requirement already satisfied: pylint>=2.0 in c:\users\dayebga\appdata\roaming\python\python37\site-packages (from pylint-django) (2.4.4)
Collecting pylint-plugin-utils>=0.5 (from pylint-django)
  Downloading https://files.pythonhosted.org/packages/f2/72/ed6a263988ea68a4236591205340630d34621f60dbfc661543d3b4216b5e/pylint_plugin_utils-0.6-py3-none-any.whl
Requirement already satisfied: astroid<2.4,>=2.3.0 in c:\users\dayebga\appdata\roaming\python\python37\site-packages (from pylint>=2.0->pylint-django) (2.3.3)
Requirement already satisfied: mccabe<0.7,>=0.6 in c:\users\dayebga\appdata\roaming\python\python37\site-packages (from pylint>=2.0->pylint-django) (0.6.1)
Requirement already satisfied: isort<5,>=4.2.5 in c:\users\dayebga\appdata\roaming\python\python37\site-packages (from pylint>=2.0->pylint-django) (4.3.21)
Requirement already satisfied: colorama; sys_platform == "win32" in c:\users\dayebga\appdata\roaming\python\python37\site-packages (from pylint>=2.0->pylint-django) (0.4.3)
Requirement already satisfied: wrapt==1.11.* in c:\users\dayebga\appdata\roaming\python\python37\site-packages (from astroid<2.4,>=2.3.0->pylint>=2.0->pylint-django) (1.11.2)
Requirement already satisfied: typed-ast<1.5,>=1.4.0; implementation_name == "cpython" and python_version < "3.8" in c:\users\dayebga\appdata\roaming\python\python37\site-packages (from astroid<2.4,>=2.3.0->pylint>=2.0->pylint-django) (1.4.0)
Requirement already satisfied: lazy-object-proxy==1.4.* in c:\users\dayebga\appdata\roaming\python\python37\site-packages (from astroid<2.4,>=2.3.0->pylint>=2.0->pylint-django) (1.4.3)
Requirement already satisfied: six~=1.12 in c:\users\dayebga\appdata\roaming\python\python37\site-packages (from astroid<2.4,>=2.3.0->pylint>=2.0->pylint-django) (1.13.0)
Installing collected packages: pylint-plugin-utils, pylint-django
Successfully installed pylint-django-2.0.13 pylint-plugin-utils-0.6
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.5
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
entrypoints==0.3
importlib-metadata==0.23
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
jupyter-client==5.3.4
jupyter-core==4.6.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==7.2.0
mysqlclient==1.4.6
parso==0.5.1
pickleshare==0.7.5
prompt-toolkit==2.0.10
pycodestyle==2.5.0
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
python-dateutil==2.8.0
pytz==2019.2
pywin32==225
pyzmq==18.1.0
six==1.13.0
sqlparse==0.3.0
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.2
zipp==0.6.0
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.5
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
entrypoints==0.3
importlib-metadata==0.23
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
jupyter-client==5.3.4
jupyter-core==4.6.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==7.2.0
mysqlclient==1.4.6
parso==0.5.1
pickleshare==0.7.5
prompt-toolkit==2.0.10
pycodestyle==2.5.0
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
python-dateutil==2.8.0
pytz==2019.2
pywin32==225
pyzmq==18.1.0
six==1.13.0
sqlparse==0.3.0
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.2
zipp==0.6.0
